# api/src/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Bon-API
  version: v0
servers:
  -
    url: 'http://localhost:8000/api/v0'
paths:
  /admin/login:
    post:
      tags:
        - Admin
      operationId: e53e573f1f1f3710234d480ededf07f6
      requestBody:
        content:
          application/json:
            schema:
              required:
                - mail
                - password
              properties:
                mail:
                  type: string
                password:
                  type: string
              type: object
              example:
                mail: monmail@gmail.com
                password: '123456789'
      responses:
        '201':
          description: 'Retourne l\''admin connecté.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '404':
          description: 'Erreur de récupération'
          content:
            application/json:
              schema:
                description: 'Message d erreur'
                type: string
              example: 'Invalid request method'
  /admin:
    get:
      tags:
        - Admin
      operationId: 89bbad68b14e55e85eb06cf5c3f28790
      responses:
        '200':
          description: 'Retourne l ensemble des admins'
          content:
            application/json:
              schema:
                description: 'Liste des admins'
                type: array
                items:
                  $ref: '#/components/schemas/Admin'
      security:
        -
          Authorization: []
    post:
      tags:
        - Admin
      operationId: e892de0644c185e81e1359a5d780aa34
      requestBody:
        content:
          application/json:
            schema:
              required:
                - mail
                - password
                - nom
              properties:
                mail:
                  type: string
                password:
                  type: string
                nom:
                  type: string
              type: object
              example:
                mail: monmail@gmail.com
                password: '123456789'
                nom: 'John Doe'
      responses:
        '201':
          description: 'Retourne l\''admin créé.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
  '/admin/{id}':
    get:
      tags:
        - Admin
      operationId: 8f5cb4a18254f7694b8608591f786bcd
      parameters:
        -
          name: id
          in: path
          description: 'ID de l\''admin'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Retourne un admin en fonction de son ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
    put:
      tags:
        - Admin
      operationId: 3b01701b92b874e4c15db51618267954
      parameters:
        -
          name: id
          in: path
          description: 'ID de l\''admin à modifier'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required:
                - name
                - prix
                - image
                - description
              properties:
                name:
                  type: string
                prix:
                  type: number
                  format: float
                image:
                  type: string
                description:
                  type: string
              type: object
              example:
                name: 'Nom du plat'
                prix: 'Prix du plat'
                image: 'Image d un plat'
                description: 'Descripion du plat'
      responses:
        '200':
          description: 'Admin modifié'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
    delete:
      tags:
        - Admin
      operationId: 5c680642d3b23069e660e5d661826369
      parameters:
        -
          name: id
          in: path
          description: 'ID de l\''admin'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Administrateur supprimé.'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
  /commande:
    get:
      tags:
        - Commande
      operationId: e8277b1f6c8de1bf2e48517b12707d41
      responses:
        '200':
          description: 'Retourne l ensemble des commandes'
          content:
            application/json:
              schema:
                description: 'Toutes les commandes'
                type: array
                items:
                  $ref: '#/components/schemas/Commande'
    post:
      tags:
        - Commande
      operationId: 624a8cc562f4d6f1dd4dba6bcc401dd6
      requestBody:
        content:
          application/json:
            schema:
              required:
                - nom
                - reference
                - prix
                - status
              properties:
                nom:
                  type: string
                reference:
                  type: string
                prix:
                  type: number
                  format: float
                status:
                  type: string
              type: object
              example:
                nom: 'John Doe'
                status: 'en cours de préparation'
                reference: RA34FS3UTS
                prix: 62.1
      responses:
        '200':
          description: 'Commande crée'
          content:
            application/json:
              schema:
                type: string
  '/commande/{id}':
    get:
      tags:
        - Commande
      operationId: 07e28f5b537b479b944f36c37a04ef7a
      parameters:
        -
          name: id
          in: path
          description: 'ID de la commande'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Retourne une commande en fonction de son ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commande'
    put:
      tags:
        - Commande
      operationId: def8ecbeddb632d63db4c951b066a010
      parameters:
        -
          name: id
          in: path
          description: 'ID de la commande à modifier'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required:
                - nom
                - reference
                - prix
                - status
              properties:
                nom:
                  type: string
                reference:
                  type: string
                prix:
                  type: number
                  format: float
                status:
                  type: string
              type: object
              example:
                nom: 'John Doe'
                status: 'en cours de préparation'
                reference: RA34FS3UTS
                prix: 62.1
      responses:
        '200':
          description: 'Commande modifiée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
    delete:
      tags:
        - Commande
      operationId: f68e2b5d8b228ff46145444cc0b5329b
      parameters:
        -
          name: id
          in: path
          description: 'ID de la commande'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Suppression réussie'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
...
components:
  schemas:
    Admin:
      properties:
        id:
          type: integer
          nullable: false
        nom:
          type: string
          nullable: false
        mail:
          type: string
          nullable: false
        password:
          type: string
          nullable: false
      type: object
    Commande:
      properties:
        id:
          type: integer
          nullable: false
        nom:
          type: string
          nullable: false
        reference:
          type: string
          nullable: false
        prix:
          type: number
          format: float
          nullable: false
        status:
          type: string
          nullable: false
        insertion:
          type: string
          nullable: true
      type: object
...
  securitySchemes:
    Authorization:
      type: http
      bearerFormat: JWT
      scheme: bearer
...